parseing update

+++ components/admin/NBAOverUnderForm.tsx
@@ -131,7 +131,7 @@ export default function NBAOverUnderForm(): JSX.Element {
     return (valueString: string) => {
       setBenchmarks({
         ...benchmarks,
-        [name]: parseInt(valueString),
+        [name]: parseFloat(valueString),
       });
     };
   };
